Remove UB and injection surface from Z, StringView, DequeList, and numeral mappings
Several extraction mappings could execute C++ undefined behavior, emit
attacker-chosen source text, or relate a proof to the wrong extracted
operation, any of which can silently invalidate the guarantees a verified
Rocq source is supposed to provide.

Z arithmetic (ZInt/ZIntBDE): Z was represented as int64_t with native signed
operations, so negation/abs of INT64_MIN, add/sub/mul overflow, and
INT64_MIN / -1 or % -1 were undefined behavior (CWE-190/CWE-682). Signed UB is
not merely a wrong number -- an optimizer may miscompile the surrounding code,
bypassing checks the source proved. Each op is now computed in the unsigned
domain (well-defined two's-complement wraparound in C++20) and cast back, with
explicit guards for the INT64_MIN / -1 and abs(INT64_MIN) edge cases. Results
are unchanged for in-range values; out-of-range cases now wrap deterministically
instead of being UB. (True unbounded-Z results still require the ZGMP bignum
mapping.)

Numeral formats (Crane Extract Numeral): the format string was stored
unvalidated and emitted verbatim as raw C++ with %n textually substituted, so a
mapping could inject arbitrary expressions/statements (CWE-94). The format is
now validated to be a numeric-literal wrapper: exactly one %n and, otherwise,
only characters that build a literal or wrapping macro/cast, rejecting the
punctuation needed for statement/string/comment/lambda injection.

StringView substr_of_string_comm: the axiom passed the end index j as
PrimString.sub's length, so for a nonzero start the proof related a length-(j-i)
view to a length-j primitive slice -- different substrings (CWE-682/CWE-345).
It now passes (j - i) as the length so both sides denote the same slice.

DequeList map/flat_map/concat: result element types were derived from
%aN.front(), which is undefined on an empty deque; inference now uses the
deque's value_type instead (CWE-476/CWE-125). Adds tests/regression/deque_empty_ops
covering all three on empty and non-empty inputs.

Also documents two inherent limitations: sv_of_string returns a non-owning view
whose validity requires the source string to outlive it (CWE-416), and the
iterative deque-backed destructor heap-allocates per element, so an allocation
failure during (noexcept) destruction can still terminate (CWE-400) -- the
call-stack-overflow risk it replaces is the property that matters and is
regression-tested.

+++ tests/regression/numeral_stress/numeral_stress.cpp
@@ -21,5 +21,6 @@ bool NumeralStress::check_range(uint64_t n) {
 
 /// 10. Mixed nat and Z in one function
 int64_t NumeralStress::mixed_arith(uint64_t n) {
-  return (static_cast<int64_t>(n) + INT64_C(100));
+  return static_cast<int64_t>(static_cast<uint64_t>(static_cast<int64_t>(n)) +
+                              static_cast<uint64_t>(INT64_C(100)));
 }
